8151430: Move setattrlist to BsdNativeDispatcher
